Always include MISC|Packaging regardless of bUseLT. Move validations into the .cs file.
