+++ OJewelry/Classes/Validations.cs
@@ -52,5 +52,66 @@ namespace OJewelry.Classes
             return ValidationResult.Success;
         }
     }
+
+    [AttributeUsage(AttributeTargets.Property)]
+    public sealed class GreaterThanZeroAttribute : ValidationAttribute
+    {
+        protected override ValidationResult IsValid(object value, ValidationContext validationContext)
+        {
+            List<CastingComponent> ccl = value as List<CastingComponent>;
+            List<StoneComponent> scl = value as List<StoneComponent>;
+            List<FindingsComponent> fcl = value as List<FindingsComponent>;
+            List<LaborComponent> lcl = value as List<LaborComponent>;
+            List<MiscComponent> mcl = value as List<MiscComponent>;
+
+            List<string> items = new List<string>();
+            if (validationContext.MemberName == "Stones" && scl != null)
+            {
+                for (int i = 0; i < scl.Count; i++)
+                {
+                    if (scl[i].Id == -1)
+                    {
+                        if (scl[i].State != SVMStateEnum.Unadded)
+                        {
+                            items.Add("[" + i + "].Id");
+                        }
+                    }
+                }
+            }
+            if (validationContext.MemberName == "Findings" && fcl != null)
+            {
+                for (int i = 0; i < fcl.Count; i++)
+                {
+                    if (fcl[i].Id == -1)
+                    {
+                        if (fcl[i].State != SVMStateEnum.Unadded)
+                        {
+                            items.Add("[" + i + "].Id");
+                        }
+                    }
+                }
+            }
+            if (validationContext.MemberName == "JewelryTypeId")
+            {
+                int? jtId = value as int?;
+                if (jtId == null)
+                {
+                    return new ValidationResult("Please choose a Jewelry Type", items);
+                }
+                else
+                {
+                    return ValidationResult.Success;
+                }
+            }
+
+            if (items.Count != 0)
+            {
+                return new ValidationResult("Validation error", items);
+            }
+            // Everything OK.
+            return ValidationResult.Success;
+        }
+    }
+
 }
