+++ OJewelry/Views/JewelryTypes/Create.cshtml
@@ -43,7 +43,7 @@
             <div class="form-group">
                 @Html.LabelFor(model => model.FinishingCost, htmlAttributes: new { @class = "control-label col-md-2" })
                 <div class="col-md-10">
-                    @Html.EditorFor(model => model.FinishingCost, new { htmlAttributes = new { @class = "form-control " + nocursor } })
+                    @Html.EditorFor(model => model.FinishingCost, new { htmlAttributes = new { @class = "form-control " } })
                     @Html.ValidationMessageFor(model => model.FinishingCost, "", new { @class = "text-danger" })
                 </div>
             </div>
